Refactor: duplicate download logic
